kallia/playground/requirements.txt
kallia==0.1.3
chainlit==2.6.2
langchain-core==0.3.69
langchain-openai==0.3.28
langchain-huggingface==0.3.0
tiktoken==0.9.0
pdfminer.six==20250506